+++ bika/lims/browser/analysisrequest/add2.py
@@ -13,12 +13,14 @@ from bika.lims import POINTS_OF_CAPTURE
 from bika.lims import api
 from bika.lims import bikaMessageFactory as _
 from bika.lims import logger
+from bika.lims.api.analysisservice import get_calculation_dependencies_for
+from bika.lims.api.analysisservice import get_service_dependencies_for
 from bika.lims.interfaces import IGetDefaultFieldValueARAddHook
 from bika.lims.utils import tmpID
 from bika.lims.utils.analysisrequest import create_analysisrequest as crar
+from bika.lims.workflow import ActionHandlerPool
 from BTrees.OOBTree import OOBTree
 from DateTime import DateTime
-from bika.lims.workflow import ActionHandlerPool
 from plone import protect
 from plone.memoize.volatile import DontCache
 from plone.memoize.volatile import cache
@@ -533,139 +535,6 @@ class AnalysisRequestAddView(BrowserView):
         analysis = api.get_object(analysis)
         return api.get_uid(analysis.getAnalysisService())
 
-    def get_calculation_dependencies_for(self, service):
-        """Calculation dependencies of this service and the calculation of each
-        dependent service (recursively).
-
-        TODO: This needs to go to bika.lims.api
-        """
-
-        def calc_dependencies_gen(service, collector=None):
-            """Generator for recursive dependency resolution.
-            """
-
-            # The UID of the service
-            service_uid = api.get_uid(service)
-
-            # maintain an internal dependency mapping
-            if collector is None:
-                collector = {}
-
-            # Stop iteration if we processed this service already
-            if service_uid in collector:
-                raise StopIteration
-
-            # Get the calculation of the service.
-            # The calculation comes either from an assigned method or the user
-            # has set a calculation manually (see content/analysisservice.py).
-            calculation = service.getCalculation()
-
-            # Stop iteration if there is no calculation
-            if not calculation:
-                raise StopIteration
-
-            # The services used in this calculation.
-            # These are the actual dependencies of the used formula.
-            dep_services = calculation.getDependentServices()
-            for dep_service in dep_services:
-                # get the UID of the dependent service
-                dep_service_uid = api.get_uid(dep_service)
-
-                # remember the dependent service
-                collector[dep_service_uid] = dep_service
-
-                # yield the dependent service
-                yield dep_service
-
-                # check the dependencies of the dependent services
-                for ddep_service in calc_dependencies_gen(dep_service,
-                                                          collector=collector):
-                    yield ddep_service
-
-        dependencies = {}
-        for dep_service in calc_dependencies_gen(service):
-            # Skip the initial (requested) service
-            if dep_service == service:
-                continue
-            uid = api.get_uid(dep_service)
-            dependencies[uid] = dep_service
-
-        return dependencies
-
-    def get_calculation_dependants_for(self, service):
-        """Calculation dependants of this service
-
-        TODO: This needs to go to bika.lims.api
-        """
-
-        def calc_dependants_gen(service, collector=None):
-            """Generator for recursive resolution of dependant sevices.
-            """
-
-            # The UID of the service
-            service_uid = api.get_uid(service)
-
-            # maintain an internal dependency mapping
-            if collector is None:
-                collector = {}
-
-            # Stop iteration if we processed this service already
-            if service_uid in collector:
-                raise StopIteration
-
-            # Get the dependant calculations of the service
-            # (calculations that use the service in their formula).
-            dep_calcs = service.getBackReferences('CalculationAnalysisService')
-            for dep_calc in dep_calcs:
-                # Get the methods linked to this calculation
-                dep_methods = dep_calc.getBackReferences('MethodCalculation')
-                for dep_method in dep_methods:
-                    # Get the services that have this method linked
-                    dep_services = dep_method.getBackReferences(
-                        'AnalysisServiceMethod')
-                    for dep_service in dep_services:
-
-                        # get the UID of the dependent service
-                        dep_service_uid = api.get_uid(dep_service)
-
-                        # skip services with a different calculation, e.g. when
-                        # the user selected a calculation manually.
-                        if dep_service.getCalculation() != dep_calc:
-                            continue
-
-                        # remember the dependent service
-                        collector[dep_service_uid] = dep_service
-
-                        # yield the dependent service
-                        yield dep_service
-
-                        # check the dependants of the dependant services
-                        for ddep_service in calc_dependants_gen(
-                                dep_service, collector=collector):
-                            yield ddep_service
-
-        dependants = {}
-        for dep_service in calc_dependants_gen(service):
-            # Skip the initial (requested) service
-            if dep_service == service:
-                continue
-            uid = api.get_uid(dep_service)
-            dependants[uid] = dep_service
-
-        return dependants
-
-    def get_service_dependencies_for(self, service):
-        """Calculate the dependencies for the given service.
-        """
-
-        dependants = self.get_calculation_dependants_for(service)
-        dependencies = self.get_calculation_dependencies_for(service)
-
-        return {
-            "dependencies": dependencies.values(),
-            "dependants": dependants.values(),
-        }
-
     def is_service_selected(self, service):
         """Checks if the given service is selected by one of the ARs.
         This is used to make the whole line visible or not.
@@ -1045,10 +914,8 @@ class ajaxAnalysisRequestAddView(AnalysisRequestAddView):
 
         })
 
-        dependencies = self.get_calculation_dependencies_for(obj).values()
+        dependencies = get_calculation_dependencies_for(obj).values()
         info["dependencies"] = map(self.get_base_info, dependencies)
-        # dependants = self.get_calculation_dependants_for(obj).values()
-        # info["dependendants"] = map(self.get_base_info, dependants)
         return info
 
     @cache(cache_key)
@@ -1562,7 +1429,7 @@ class ajaxAnalysisRequestAddView(AnalysisRequestAddView):
             #  DEPENDENCIES
             for uid, obj in _services.iteritems():
                 # get the dependencies of this service
-                deps = self.get_service_dependencies_for(obj)
+                deps = get_service_dependencies_for(obj)
 
                 # check for unmet dependencies
                 for dep in deps["dependencies"]:
@@ -1578,8 +1445,6 @@ class ajaxAnalysisRequestAddView(AnalysisRequestAddView):
                 service_metadata[uid].update({
                     "dependencies": map(
                         self.get_base_info, deps["dependencies"]),
-                    "dependants": map(
-                        self.get_base_info, deps["dependants"]),
                 })
 
             # Each key `n` (1,2,3...) contains the form data for one AR Add